src/requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
executing==1.0.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==4.21.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
